Update variable declarations to use ES6 style
Updated variable and function declarations to use let or const. const
was used in cases where a function was being declared or the variable
appeared to be immutable.  let was used for all others.

I would reccomend further work to change function declaration to use the
syntax:
  "function foo(arg1, arg2) { }"

And the main function should be exported by:
  "export default function Ringme() {
   ...
   }"